+++ CHANGELOG.rst
@@ -117,6 +117,7 @@ Added
 
 - Labelling of atomic orbital choices in output.
 
+- Halogen X correction.
 
 Changed
 -------